024-tables_format/requirements.txt
openpyxl==2.6.1
pandas==0.24.2
plotly==3.7.1
pyarrow==0.12.1
pyyaml==5.1
tqdm==4.31.1
xlrd==1.2.0
